QProgressbar without text does not progress smoothly
The jerk with QProgressBar occurs when setTextVisible(false) is used.
Condition exists to handle repaint when text visibility is true but not
handled when text visibility is false.
Usage of auto leads to type confusions.

Add the logic to repaint for every 1 pixel change in the value when
text visibility is false.
Avoid using auto when type is obvious.

Fixes: QTBUG-45048
Change-Id: I0fdc62c7aabdaef52bc4bdc20e82141a4a7d29d5
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
(cherry picked from commit 913b98d4120b56487a14eb96b5e5252fdd939603)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>