CMake: Remove deployment file check in android test runner args query
The QT_ANDROID_DEPLOYMENT_SETTINGS_FILE was originally queried and
checked as part of the following commit
080f9ad160ef7422b2bab884c7ccc950d35b3a0a
because it was meant to be passed to the test runner executable.

The deployment file argument stopped being passed to the test runner
as part of 29b17fa335388c9b93f70c29b2398cf2fee65785 whereas the
check was left in the code.

This check causes problems when trying to allow running deferred
android finalizers for tests, because the deployment file is set only
in the finalizer, but the check for the deployment file existence is
done earlier when trying to get the test runner arguments.

Remove the query and the check for QT_ANDROID_DEPLOYMENT_SETTINGS_FILE
because they prevent the above use case, and also don't serve any
purpose in the given function scope.

Task-number: QTBUG-93625
Task-number: QTBUG-112212
Change-Id: I6e620e452359c94cc7896a98ac5872221f6bb9a2
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
(cherry picked from commit 412e08106c050cd6b3f6bdf4d9303850c32f6232)
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>