CMake: Fix install location of dSYM bundle for Qt apps
Previously we used the TARGET_FILE genex as the destination where
to place the extracted dSYM bundle of an app in the build folder. For
Qt apps that have the MACOSX_BUNDLE property set, this is incorrect,
because it placed the dSYM bundle inside the app bundle.

We then manually install the dSYM bundle to the correct location.

Because the build folder app bundle still contains the dSYM
bundle, and CMake copies the entire app bundle dir when installing the
app, this resulted in installing the dSYM bundle twice.

Aside from waste of space, this is also against the Apple guidelines,
and leads to certain issues with code-signing.

Check the MACOSX_BUNDLE property to determine if the executable is a
bundle, in which case treat it the same way we treat framework
bundles, by placing the dSYM bundle next to the app bundle.

Change-Id: I7fb8fe49e0b26d0df8a32fc82530bbee32b51382
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit a60d42d873e67281e5e30bcaa3daa2f8edffe7eb)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>